added the ability to edit the monitors

+++ src/main/java/by/gstu/computerdetails/entity/Monitor.java
@@ -74,7 +74,7 @@ public class Monitor extends BaseEntity {
     }
 
     public void setPrice(BigDecimal price) {
-        long intPrice = price.longValue();
+        double intPrice = price.doubleValue();
         if (intPrice < 0) {
             throw new IllegalArgumentException("Price less than zero.");
         }
